# Performance optimizations