--- a/src/main/java/io/r2dbc/h2/CommandUtil.java
+++ b/src/main/java/io/r2dbc/h2/CommandUtil.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2019 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.r2dbc.h2;
+
+import org.h2.command.Command;
+import org.h2.command.CommandInterface;
+
+/**
+ * @author Mark Paluch
+ */
+final class CommandUtil {
+
+    /**
+     * Releases a {@link Command} for reuse.
+     *
+     * @param command
+     */
+    static void clearForReuse(CommandInterface command) {
+        if (command instanceof Command) {
+            ((Command) command).setCanReuse(true);
+        }
+    }
+
+    private CommandUtil() {
+    }
+}