Reuse parsed commands by setting the reuse flag after executing a command.

Removal of deferred invocations in Client as facades (Connection, Statement) already defer execution. Usage of imperative API avoids excessive Publisher creation and code paths through Subscriptions.

Revise metadata handling. Delay `H2ColumnMetadata.getJavaType()` retrieval to not determine the outcome upfront. Perform `String`-based column lookups using exact lookup first and then fall back to case-insensitive lookup. Merge Collection of column names into H2RowMetadata to avoid object instantiations.

Benchmarks:

```
Before:
Benchmark                               Mode  Cnt        Score   Error  Units
StatementBenchmarks.parametrizedJdbc   thrpt    2  2144074,405          ops/s
StatementBenchmarks.parametrizedR2dbc  thrpt    2   179761,369          ops/s
StatementBenchmarks.simpleJdbc         thrpt    2  2392871,130          ops/s
StatementBenchmarks.simpleR2dbc        thrpt    2   198372,569          ops/s


After:
Benchmark                               Mode  Cnt        Score   Error  Units
StatementBenchmarks.parametrizedJdbc   thrpt    2  2559550,182          ops/s
StatementBenchmarks.parametrizedR2dbc  thrpt    2   663763,110          ops/s
StatementBenchmarks.simpleJdbc         thrpt    2  2385740,695          ops/s
StatementBenchmarks.simpleR2dbc        thrpt    2   600211,907          ops/s
```